Reorganize into module trees and fix every clippy lint (#13)
* Split renderer, apu and bus into modules; drop dead code

Pure reorganization plus dead-code removal: no behavior changes, and the
same tests pass before and after (328 core + 8 real-ROM boot + 32 frontend).
Code was moved verbatim, so the doc comments explaining past hardware bugs
travel with the code they describe.

Dead code removed:
- `oxidesfc-core/src/io.rs` (872 lines). CLAUDE.md already flagged it as an
  unused and known-incorrect register map that must not be confused with the
  real dispatch in `bus.rs`; nothing but `lib.rs` referenced it. Its 18 tests
  went with it, which is why the core count drops from 346 to 328 -- they
  pinned the behavior of a module that never ran.
- `input/keyboard.rs`, a `KeyboardState` that was never constructed (real
  keyboard handling is entirely frontend DOM events).
- The `windows` crate's `Gaming_Input` dependency, declared but never
  referenced anywhere in `src-tauri`.
- `Adsr::reset`, `Apu::spc700`'s public form, and the SPC700 PSW's `u`/`d`
  fields -- 6502 leftovers never read by anything. The real SPC700 PSW is
  exactly N V P B H I Z C, all eight of which already had their own field,
  and keeping `d` invited the assumption that the SPC700 has a 6502-style
  decimal flag (it does not; `p` is direct-page select).

File layout, largest production file per module in parentheses:
- `renderer/` (compose 411) -- compose, background, sprites, mode7, window,
  color, tile, and tests split to match.
- `apu/` (opcodes 859) -- spc700 + its opcode dispatch, dsp, voice, brr,
  envelope, and the `Apu` wrapper.
- `bus/` (write 528) -- core, read, write, ports, transfer, timing, state.

`apu/opcodes.rs` and `bus/{read,write}.rs` stay long on purpose: each is one
exhaustive `match`. The SPC700 dispatch in particular has no wildcard arm, so
the compiler itself proves all 256 encodings are handled -- splitting the arms
into per-range helpers would give each one an unreachable catch-all and throw
that guarantee away.

Visibility was tightened rather than widened while moving things: items that
were file-private are now `pub(super)`/`pub(in crate::apu)` -- the same
effective scope -- and `Spc700` went from `pub` to module-internal, since
`lib.rs` only ever exported `Apu`. `Apu::spc700` is now `#[cfg(test)]`, which
is what it always was in practice.

Also corrected a stale doc comment claiming the SPC700 implements "exactly
the opcode subset used by the IPL boot ROM" -- all 256 are implemented and
pinned by a test.


* Split cpu, ppu, cartridge, oam and dma into modules

Continues the reorganization; still no behavior changes, same tests passing.

`cpu.rs` was the largest file in the tree at 6123 lines. Its single
`impl Cpu` is now split across files by instruction family, so an opcode sits
next to its relatives: `core` (step loop, memory access, width/flag
bookkeeping), `dispatch` (the 256-arm match), `addressing`, then
`load_store`, `add_sub`, `compare`, `incdec`, `logic`, `bittest`, `shift`,
`branch`, `stack`, `transfer`, `flags`, and `interrupts`. All 309 methods were
classified by name so none could be silently dropped, and the tests are
grouped the same way.

`ppu.rs` splits `PpuRegisters` into its own `registers` module, which also
documents why it exists as data separate from `Ppu`: the renderer needs a copy
per scanline, since games rewrite scroll/windows/color-math/mode-7 mid-frame.
Its old `///` header (which described the module but was attached to the next
item) became a real `//!` module doc.

`cartridge`, `oam` and `dma` keep their production code as-is and move their
inline test modules into `<name>/tests.rs`.

No file in the core is now over 900 lines, and the only ones over 500 are the
two exhaustive opcode matches, the bus register dispatch, and four test
modules.


* Split the frontend modules and fix every clippy lint

Completes the reorganization and the Rust best-practices pass. The workspace
now builds with zero warnings in debug, in release, and under
`cargo clippy --workspace --all-targets`, with the same tests passing
throughout (328 core + 8 real-ROM boot + 32 frontend, real ROMs included).

Frontend layout:
- `emulation/controller.rs` split into `snes.rs` (the `Cpu` + `SystemBus`
  composition and the region/frame-rate tables) and the controller proper,
  with its two test modules as siblings.
- `commands/library.rs` split into `library/{scan,store}` plus the command
  handlers. Note the constraint discovered here and written into CLAUDE.md:
  `invoke_handler!` resolves a companion macro from the command's own module
  path, so `#[tauri::command]` fns must stay in the module named there.

Encapsulation rather than public fields: the real-ROM tests used to reach
through `controller.snes.bus` and `snes.cpu.pc` directly. `Snes` now offers
`video_mode()` for the frame pacer plus `#[cfg(test)]` accessors
(`bus_mut`, `force_halt`, `program_counter`, `pc_and_accumulator`), so no
field had to be opened up to make the split compile.

Clippy, all fixed rather than allow-ed:
- `too_many_arguments` (6 sites). The renderer's draw calls took 10-11
  positional parameters; they now take a `Target` (the color buffer paired
  with its parallel layer-id buffer -- these must stay in lockstep, so they
  travel together), a `Frame` (the read-only VRAM/CGRAM/OAM/registers), and a
  `Band` (the y0..y1 range). `Voice::sample` took 13 and now takes a named
  `VoiceConfig`, which matters here specifically: an early version of this
  code had the register-to-field mapping offset by one slot, so `srcn` read
  VOL(L) and every voice played the wrong sample. Named fields make that
  class of mistake visible at the call site.
- `field_reassign_with_default` (25 sites) -> struct-update syntax, comments
  preserved per field.
- `needless_range_loop`, `unnecessary_cast`, `assign_op_pattern`,
  `identity_op`, `manual_rotate`, `manual_is_multiple_of`, `derivable_impls`,
  `ptr_arg` (`&PathBuf` -> `&Path`), `wrong_self_convention`
  (`Psw::to_byte` takes `self`, being `Copy`), `upper_case_acronyms`
  (`Country::USA` -> `Usa`, with `#[serde(rename)]` and its `as_str()` value
  unchanged so nothing user-visible moves), and the doc-comment lints -- two
  of which were `///` blocks describing a module while attached to a struct,
  now real `//!` module docs.
- `erasing_op`: `0 * SCREEN_WIDTH` written to mean "row 0".

Also removed a `mut` that only release builds warned about, by shadowing the
Tauri builder under `cfg(debug_assertions)` instead of reassigning it.

CLAUDE.md's architecture section described the old single-file layout, the
deleted `io.rs`/`keyboard.rs`, and the removed `windows` dependency; it now
describes the module tree, says which files are deliberately long and why,
and points at `set_video_mode` as the correct way to select NTSC/PAL.

Verified beyond the test suite: a real SMW frame rendered after the refactor
is statistically identical to one rendered before it (42 distinct colors,
mean luma 159.2) and was inspected by eye.


---------

+++ OxideSFC/oxidesfc-core/tests/real_rom_boot.rs
@@ -144,11 +144,11 @@ fn cpu_executes_genuine_boot_instructions_until_a_known_gap() {
                 // own genuine IPL boot ROM (see oxidesfc_core::apu) -- it
                 // must be ticked forward in lockstep with CPU cycles, same
                 // as a real frontend driving the emulator would.
-                bus.tick_apu(cycles as u32);
+                bus.tick_apu(cycles);
                 // Same for the PPU: real hardware fires NMI once per frame
                 // at vblank entry, which is what actually unblocks SMW's
                 // boot code (it spins on $4210/RDNMI waiting for vblank).
-                bus.tick_ppu(cycles as u32);
+                bus.tick_ppu(cycles);
                 dispatch_interrupts(&mut cpu, &mut bus);
                 steps_executed += 1;
             }
@@ -221,8 +221,8 @@ fn cpu_sustains_long_run_with_real_frame_and_state_progression() {
         unique_pcs.insert(((cpu.pb as u32) << 16) | (cpu.pc as u32));
         match cpu.step(&mut bus) {
             Ok(cycles) => {
-                bus.tick_apu(cycles as u32);
-                bus.tick_ppu(cycles as u32);
+                bus.tick_apu(cycles);
+                bus.tick_ppu(cycles);
                 if bus.take_pending_nmi() {
                     nmi_count += 1;
                     cpu.nmi(&mut bus).expect("NMI delivery must not fault mid-run");
@@ -330,8 +330,8 @@ fn real_dma_transfers_populate_vram_cgram_oam_with_genuine_cartridge_data() {
     for _ in 0..MAX_STEPS {
         match cpu.step(&mut bus) {
             Ok(cycles) => {
-                bus.tick_apu(cycles as u32);
-                bus.tick_ppu(cycles as u32);
+                bus.tick_apu(cycles);
+                bus.tick_ppu(cycles);
                 dispatch_interrupts(&mut cpu, &mut bus);
                 steps_executed += 1;
             }
@@ -397,8 +397,8 @@ fn real_rom_eventually_renders_a_visible_non_black_frame() {
     for _ in 0..MAX_STEPS {
         match cpu.step(&mut bus) {
             Ok(cycles) => {
-                bus.tick_apu(cycles as u32);
-                bus.tick_ppu(cycles as u32);
+                bus.tick_apu(cycles);
+                bus.tick_ppu(cycles);
                 dispatch_interrupts(&mut cpu, &mut bus);
                 steps_executed += 1;
             }
@@ -478,8 +478,8 @@ fn gamemode_advances_past_the_former_bank_ef_escape_point() {
 
         match cpu.step(&mut bus) {
             Ok(cycles) => {
-                bus.tick_apu(cycles as u32);
-                bus.tick_ppu(cycles as u32);
+                bus.tick_apu(cycles);
+                bus.tick_ppu(cycles);
                 dispatch_interrupts(&mut cpu, &mut bus);
                 steps_executed += 1;
             }
